internal: move priorities to base-core
