+++ generators/base-application/priorities.ts
@@ -18,7 +18,7 @@
  */
 
 import type { Priority } from 'yeoman-generator';
-import { PRIORITY_NAMES as PRIORITY_NAMES_BASE, QUEUES as QUEUES_BASE, QUEUE_PREFIX } from '../base/priorities.js';
+import { PRIORITY_NAMES as PRIORITY_NAMES_BASE, QUEUES as QUEUES_BASE, QUEUE_PREFIX } from '../base-core/priorities.ts';
 import { BOOTSTRAP_APPLICATION } from '../base-simple-application/priorities.js';
 
 const { DEFAULT, TRANSFORM, MULTISTEP_TRANSFORM } = PRIORITY_NAMES_BASE;
