+++ generators/base-core/priorities.ts
