+++ generators/base-simple-application/priorities.ts
@@ -1,6 +1,6 @@
 import type { Priority } from 'yeoman-generator';
 import type CoreGenerator from '../base-core/generator.js';
-import { PRIORITY_NAMES, QUEUE_PREFIX } from '../base/priorities.js';
+import { PRIORITY_NAMES, QUEUE_PREFIX } from '../base-core/priorities.ts';
 
 export const BOOTSTRAP_APPLICATION = 'bootstrapApplication';
 export const BOOTSTRAP_APPLICATION_QUEUE = `${QUEUE_PREFIX}${BOOTSTRAP_APPLICATION}`;
