+++ generators/base/generator.ts
@@ -37,7 +37,7 @@ import { packageNameToNamespace } from '../../lib/utils/index.js';
 import { formatDateForChangelog } from '../base/support/index.js';
 import type { TaskTypes as BaseTaskTypes } from './tasks.js';
 import { mergeBlueprints, normalizeBlueprintName, parseBlueprints } from './internal/index.js';
-import { PRIORITY_NAMES } from './priorities.js';
+import { PRIORITY_NAMES } from '../base-core/priorities.ts';
 import {
   CONTEXT_DATA_BLUEPRINTS_TO_COMPOSE,
   CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP,
