+++ generators/gradle/generator.ts
@@ -23,7 +23,7 @@ import BaseApplicationGenerator from '../base-application/index.js';
 
 import { GRADLE_BUILD_SRC_DIR } from '../generator-constants.js';
 import { mutateData } from '../../lib/utils/index.js';
-import { QUEUES } from '../base/priorities.js';
+import { QUEUES } from '../base-core/priorities.ts';
 import files from './files.js';
 import { GRADLE } from './constants.js';
 import cleanupOldServerFilesTask from './cleanup.js';
